KnowledgeB/requirements.txt
flask==3.0.0
flask-cors==4.0.0
mem0ai
openai
python-dotenv
qdrant-client
PyPDF2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1
requests==2.31.0